Add Absorb and adjust misc buff names
