+++ relive-simulator-core/src/commonMain/kotlin/xyz/qwewqa/relive/simulator/core/presets/dress/boss/seesaw/sakurataisen/SakuraTaisenStageGirlMahiru.kt
@@ -67,7 +67,7 @@ val sakuraTaisenStageGirlMahiruBoss = ActorLoadout(
             ActType.Act3("Accuracy Slash", 2) {
                 targetSelf().act {
                     applyBuff(
-                        PerfectAim,
+                        PerfectAimBuff,
                         turns = 1,
                     )
                 }
