+++ relive-simulator-core/src/commonMain/kotlin/xyz/qwewqa/relive/simulator/core/presets/dress/boss/tr/tr10/TR10Hellsing.kt
@@ -108,7 +108,7 @@ val tr10HellsingMichiru = ActorLoadout(
                         turns = 3,
                     )
                     applyBuff(
-                        PerfectAim,
+                        PerfectAimBuff,
                         turns = 1,
                     )
                 }
@@ -127,7 +127,7 @@ val tr10HellsingMichiru = ActorLoadout(
                         turns = 3,
                     )
                     applyBuff(
-                        PerfectAim,
+                        PerfectAimBuff,
                         turns = 1,
                     )
                 }
