+++ relive-simulator-core/src/commonMain/kotlin/xyz/qwewqa/relive/simulator/core/presets/dress/boss/tr/tr10/TR10Vampire.kt
@@ -73,7 +73,7 @@ val tr10VampireShiori = ActorLoadout(
             ActType.Act6("Perfect Aim Concerto", 2) {
                 targetSelf().act {
                     applyBuff(
-                        PerfectAim,
+                        PerfectAimBuff,
                         turns = 1,
                     )
                 }
