+++ relive-simulator-core/src/commonMain/kotlin/xyz/qwewqa/relive/simulator/core/presets/dress/boss/tr/tr11/TR11CheerTsukasa.kt
@@ -73,7 +73,7 @@ val tr11CheerTsukasa = ActorLoadout(
             ActType.Act6("Perfect Aim Flurry", 2) {
                 targetSelf().act {
                     applyBuff(
-                        PerfectAim,
+                        PerfectAimBuff,
                         turns = 1,
                     )
                 }
