+++ relive-simulator-core/src/commonMain/kotlin/xyz/qwewqa/relive/simulator/core/presets/dress/boss/tr/tr12/TR12VampireShiori.kt
@@ -73,7 +73,7 @@ val tr12VampireShiori = ActorLoadout(
             ActType.Act6("Perfect Aim Concerto", 2) {
                 targetSelf().act {
                     applyBuff(
-                        PerfectAim,
+                        PerfectAimBuff,
                         turns = 1,
                     )
                 }
